backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
pydantic-settings
passlib[bcrypt]
python-multipart
PyJWT
httpx
